Move "NativeKafkaSerdeProvider" out of API
