+++ serde/src/test/java/org/creekservice/internal/kafka/serde/provider/NativeKafkaSerdeProviderTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.creekservice.api.kafka.serde;
+package org.creekservice.internal.kafka.serde.provider;
 
 import static org.creekservice.api.kafka.metadata.SerializationFormat.serializationFormat;
 import static org.hamcrest.MatcherAssert.assertThat;
